requirements.txt
# Core packages
python-dotenv
fastapi
hdbcli
langchain
generative-ai-hub-sdk[all]
mcp[stdio]
httpx
uvicorn
